pandoc_browse_extension {pandoc}	R Documentation

Open Pandoc's documentation about an extension

Description

Open Pandoc's documentation about an extension

Usage

pandoc_browse_extension(extension = NULL)

Arguments

extension
One of the supported extension. See pandoc_list_extensions(). As the Pandoc MANUAL only concerns the last released Pandoc's version, if the URL is incorrect this could mean the extensions has changed.

Value

Open the webpage at the place regarding the required extension.

Examples

pandoc_browse_extension()
pandoc_browse_extension("auto_identifiers")

[Package pandoc version 0.2.0 Index]