pandoc_list_abbreviations {pandoc}	R Documentation

List system default abbreviations

Description

Pandoc uses this list in the Markdown reader. Strings found in this list will be followed by a non-breaking space, and the period will not produce sentence-ending space in formats like LaTeX. The strings may not contain spaces.

Usage

pandoc_list_abbreviations(version = "default")

Arguments

version
Version to use. Default will be the "default" version. Other possible value are

A version number e.g "2.14.1"

The nightly version called "nightly"

The latest installed version with "latest"

Pandoc binary shipped with RStudio IDE with "rstudio"

Pandoc binary found in PATH with "system"

Details

This correspond to the option --abbreviations as CLI flag.

Value

a character version of system default abbreviation known by Pandoc

Examples

pandoc_list_abbreviations()

# check abbreviations available in a specific Pandoc's version
pandoc_list_abbreviations("2.11.4")

[Package pandoc version 0.2.0 Index]