pandoc_list_extensions {pandoc}	R Documentation

List supported extensions for a format

Description

Pandoc has a system of extensions to activate or deactivate some features. Each format have a set of activated by default extensions and other supported extensions than can be activated.

Usage

pandoc_list_extensions(format = "markdown", version = "default")

Arguments

format
One for the supported input or output formats. See pandoc_list_formats(). It corresponds to call

version
Version to use. Default will be the "default" version. Other possible value are

A version number e.g "2.14.1"

The nightly version called "nightly"

The latest installed version with "latest"

Pandoc binary shipped with RStudio IDE with "rstudio"

Pandoc binary found in PATH with "system"

Details

All the extensions for the last Pandoc version released are available in https://pandoc.org/MANUAL.html.

Value

a data.frame (or a tibble if available) with 3 columns:

format: One of the Pandoc format

extensions : name of the extensions

default: Is the extensions activated by default or not ?

Examples

pandoc_list_extensions("markdown")
pandoc_list_extensions("gfm")

# target a specific version
pandoc_list_extensions("html", version = "system")

[Package pandoc version 0.2.0 Index]