pandoc_list_highlight_style {pandoc}	R Documentation

List supported styles for Pandoc syntax highlighting

Description

Pandoc includes a highlighter which offer a styling mechanism to specify the coloring style to be used in highlighted source code. This function returns the supported values which can be specify at pandoc command line using the --highlight-style= flag.

Usage

pandoc_list_highlight_style(version = "default")

Arguments

version
Version to use. Default will be the "default" version. Other possible value are

A version number e.g "2.14.1"

The nightly version called "nightly"

The latest installed version with "latest"

Pandoc binary shipped with RStudio IDE with "rstudio"

Pandoc binary found in PATH with "system"

Value

a character vector of supported highlighting style name to use.

Examples

pandoc_list_highlight_style()

# check style available in a specific Pandoc's version
pandoc_list_highlight_style("2.11.4")

[Package pandoc version 0.2.0 Index]