pandoc_locate {pandoc}	R Documentation

Locate a specific Pandoc version installed with this package

Description

This package helps install and manage Pandoc binaries in a specific folder. This function helps with finding the path to those specific versions of Pandoc. See pandoc_bin() for another way of getting paths to pandoc binaries

Usage

pandoc_locate(version = "default")

Arguments

version
Version to use. Default will be the "default" version. Other possible value are

A version number e.g "2.14.1"

The nightly version called "nightly"

The latest installed version with "latest"

Pandoc binary shipped with RStudio IDE with "rstudio"

Pandoc binary found in PATH with "system"

Value

Path of Pandoc binaries root folder if version is available.

See Also

pandoc_install()

Examples

# where is the default active version located ?
pandoc_locate()

pandoc::pandoc_is_installed("2.11.4")
# where is a specific installed version located
pandoc_locate("2.11.4")
# return root folder of installed versions
pandoc_locate(NULL)

[Package pandoc version 0.2.0 Index]