pandoc_run {pandoc}	R Documentation

Run the pandoc binary from R

Description

This function is a thin wrapper around the pandoc binary and allow to pass any arguments supported by the Pandoc binary.

Usage

pandoc_run(args, version = "default")

Arguments

args
Character vector, arguments to the pandoc CLI command

version
Version to use. Default will be the "default" version. Other possible value are

A version number e.g "2.14.1"

The nightly version called "nightly"

The latest installed version with "latest"

Pandoc binary shipped with RStudio IDE with "rstudio"

Pandoc binary found in PATH with "system"

Value

The output of running pandoc binary

Examples

# Run any command line argument (prefer `pandoc_convert()` for conversion )
pandoc::pandoc_run(c("--version"))

pandoc::pandoc_run(c("--list-input-formats"), version = "system")

[Package pandoc version 0.2.0 Index]