rstudio_pandoc {pandoc}	R Documentation

Retrieve path and version of Pandoc shipped with RStudio

Description

RStudio IDE ships with a pandoc binary. The path is stored in RSTUDIO_PANDOC environment variable. Theses function are helper to easily use this specific version.

Usage

pandoc_rstudio_version()

pandoc_rstudio_bin()

Value

pandoc_rstudio_version() returns the version number for pandoc binary used by RStudio IDE as a base::numeric_version() object.

pandoc_system_bin() returns absolute path to the pandoc binary used by RStudio IDE.

See Also

pandoc_version(), pandoc_bin()

Examples

pandoc_rstudio_bin()

[Package pandoc version 0.2.0 Index]