system_pandoc {pandoc}	R Documentation

Retrieve path and version of Pandoc found on the system PATH

Description

Pandoc can also be installed on a system and available through the PATH. Theses function are helper to easily use this specific version.

Usage

pandoc_system_version()

pandoc_system_bin()

Value

pandoc_system_version() returns the version number for pandoc binary found in PATH as a base::numeric_version() object.

pandoc_system_bin() returns absolute path to the pandoc binary found in PATH.

See Also

pandoc_version(), pandoc_bin()

Examples

pandoc_system_bin()

[Package pandoc version 0.2.0 Index]