Check Pandoc versions already installed

pandoc_installed_versions() lists all versions already installed

pandoc_installed_latest() returns the most recent installed version

pandoc_is_installed() allows to check for a specific installed version

Usage

pandoc_installed_versions()

pandoc_installed_latest()

pandoc_is_installed(version, error = FALSE, ask = FALSE)

Arguments

version

This can be either:

"latest" for the latest release

A version number (e.g "2.11.4") for a specific version

"nightly" for the last pandoc development built daily

error

if TRUE an error will be raised if the result is FALSE

ask

if TRUE, the user will be prompt in an interactive

Value

A character vector of installed versions or a logical for pandoc_is_installed(). It will return NULL is no versions are installed.

For pandoc_is_installed(), TRUE if only the required version is installed. If FALSE and ask is TRUE, the user will be prompt for installing the version.

Examples

pandoc_installed_versions()
#>  [1] "nightly"  "3.1.6.2"  "3.1.6.1"  "3.1.6"    "3.1.5"    "3.1.4"   
#>  [7] "3.1.3"    "3.1.2"    "3.1.1"    "3.1"      "3.0.1"    "3.0"     
#> [13] "2.19.2"   "2.19.1"   "2.19"     "2.18"     "2.17.1.1" "2.17.1"  
#> [19] "2.17.0.1" "2.17"     "2.16.2"   "2.16.1"   "2.16"     "2.15"    
#> [25] "2.14.2"   "2.14.1"   "2.14.0.3" "2.14.0.2" "2.14.0.1" "2.14"    
#> [31] "2.13"     "2.12"     "2.11.4"   "2.11.3.2" "2.11.3.1" "2.11.3"  
#> [37] "2.11.2"   "2.11.1.1" "2.11.1"   "2.11.0.4" "2.11.0.2" "2.11.0.1"
#> [43] "2.11"     "2.10.1"   "2.10"     "2.9.2.1"  "2.9.2"    "2.9.1.1" 
#> [49] "2.9.1"    "2.9"      "2.8.1"    "2.8.0.1"  "2.8"      "2.7.3"   
#> [55] "2.7.2"    "2.7.1"    "2.7"      "2.6"      "2.5"      "2.4"     
#> [61] "2.3.1"    "2.3"      "2.2.3.2"  "2.2.3.1"  "2.2.2.1"  "2.2.2"   
#> [67] "2.2.1"    "2.2"      "2.1.3"    "2.1.2"    "2.1.1"    "2.1"     
#> [73] "2.0.6"    "2.0.5"    "2.0.4"    "2.0.3"   
pandoc_installed_latest()
#> [1] "3.1.6.2"
pandoc_is_installed("2.19.2")
#> [1] TRUE
pandoc_installed_latest()
#> [1] "3.1.6.2"